Mail Configuration

+++ db/seeds.rb
@@ -8,9 +8,12 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 
-Rent.delete_all
-User.delete_all
-Book.delete_all
-User.create(first_name: "alex", last_name: "falcon", email: "alex@hotmail.com", password: "123456789", confirmed_at: Time.zone.now)
+User.delete_all if (User.count != 0)
+Book.delete_all if (Book.count != 0)
+Rent.delete_all if (Rent.count != 0)
+
+User.create(first_name: "alex", last_name: "falcon", email: "alex@hotmail.com", password: "123456789", locale: "en", confirmed_at: Time.zone.now)
 Book.create(gender: "Metafiction", author: "Carlene Prosacco", image: "http://ziemann.net/andrea.cain", title: "Frequent Hearses", editor: "Tate Publishing \u0026 Enterprises", year: "2019-08-13")
-Rent.create!(user_id: 1, book_id: 1, start_date: Date.today, end_date: Date.today)
+user = User.last
+book = Book.last
+Rent.create!(user_id: user.id, book_id: book.id, start_date: Date.today, end_date: Date.today)
